+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Implemented websockets
 - Less logging at Info level, enable debug level for more details
+- For AP, the number of wireless clients is always shown, regardless of config option for map
 
 ### Fixed
 - Labels in lovelace editor not showing up